feat(a11y): enable underline to links by default
please note that this is an experimental change. i personally have some
sort of trouble actually reading the text now, so i might change it to
not have the underline, but instead an "external site" or "link" icon at
the upper right.

please do open an issue/submit a pr if you have better ideas!!